Signup Page Updates
I created a SignupForm for consistency in styling and CSS, edited the Signup to implement the SignupForm and wrapped the if statements for safety. I also Added confirmPassword for safety of user.
Included an update in the theme.csss so that errors could appear as errors and in the color red for aesthetics/user understanding.